book-recommender-bot/requirements.txt
aiogram==3.3.0
openai==1.30.1
SQLAlchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
